requirements.txt
crewai
crewai_tools
python-dotenv
feedparser
beautifulsoup4
requests
scikit-learn
pytest